fix and test for Bugzilla Bug 42993 	  	Language regression, or possible language improvement?
The problem was caused by moving name binding in pointcut declarations to 
happen before declare parents are evaluated.  Because of this, the 
compiler doesn't know that ContainerDescriptor isa Key when resolving 
the ContainerLoader.containerLoads reference.

The change in ordering was made to fix a bug reported in declare error 
and declare soft whose pcds where being evaluated before name binding 
had happened in the pointcut declarations.  Unfortunately, declare error 
and declare soft are concretized at the same time as declare parents 
(and all other declares ;-), so this move also led to the regression 
noted above.